Feat : 뷰 및 뷰 리졸버 기능 구현
클라이언트에게 응답할 때, 2가지 경우가 있다.
1. 리다이렉트
2. 그 외
따라서 2가지 경우에 대해 각각 뷰를 만들어주었다.

뷰 리졸버는 요청에 맞는 뷰를 찾는 기능으로 구현하였다.
